+++ lib/skylight/railtie.rb
@@ -34,7 +34,10 @@ module Skylight
       end
 
       config = Config.load(path, Rails.env.to_s, ENV)
-      config.logger = Rails.logger
+      config['root'] = Rails.root
+
+      configure_logging(config, app)
+
       config['agent.sockfile_path'] = tmp
       config['normalizers.render.view_paths'] = app.config.paths["app/views"].existent
       config.validate!
@@ -43,6 +46,28 @@ module Skylight
     rescue ConfigError => e
       puts "[SKYLIGHT] #{e.message}; disabling Skylight agent"
       nil
+    end
+
+    def configure_logging(config, app)
+      if logger = app.config.skylight.logger
+        config.logger = logger
+      else
+        # Configure the log file destination
+        if log_file = app.config.skylight.log_file
+          config['log_file'] = log_file
+        elsif !config.key?('log_file')
+          config['log_file'] = File.join(Rails.root, 'log/skylight.log')
+        end
+
+        # Configure the log level
+        if level = app.config.skylight.log_level
+          config['log_level'] = level
+        elsif !config.key?('log_level')
+          if level = app.config.log_level
+            config['log_level'] = level
+          end
+        end
+      end
     end
 
     def config_path(app)