+++ spec/unit/custom_instrumentation_spec.rb
@@ -95,6 +95,97 @@ describe Skylight::Instrumenter, :http do
         duration:   1_000)
     end
 
+    class MyClass
+      include Skylight::Helpers
+
+      instrument_method
+      def one
+        yield if block_given?
+      end
+
+      def two
+        yield if block_given?
+      end
+
+      def three
+        yield if block_given?
+      end
+
+      instrument_method category: "app.winning", title: "Win"
+      def custom
+        yield if block_given?
+      end
+
+      instrument_method :three
+
+      instrument_method
+      def self.singleton_method
+        yield if block_given?
+      end
+
+    end
+
+    it 'tracks instrumented methods using the helper' do
+      Skylight.trace 'Testin', 'app.rack.request' do |t|
+        inst = MyClass.new
+
+        clock.skip 0.1
+        ret = inst.one { clock.skip 0.1; :one }
+        ret.should == :one
+
+        clock.skip 0.1
+        inst.two { clock.skip 0.1 }
+
+        clock.skip 0.1
+        inst.three { clock.skip 0.1 }
+
+        clock.skip 0.1
+        inst.custom { clock.skip 0.1 }
+
+        clock.skip 0.1
+        MyClass.singleton_method { clock.skip 0.1 }
+      end
+
+      clock.unfreeze
+      server.wait
+
+      server.reports[0].should have(1).endpoints
+
+      ep = server.reports[0].endpoints[0]
+      ep.name.should == 'Testin'
+      ep.should have(1).traces
+
+      t = ep.traces[0]
+      t.should have(5).spans
+
+      t.spans[0].should == span(
+        event:      event('app.method', 'MyClass#one'),
+        started_at: 1_000,
+        duration:   1_000)
+
+      t.spans[1].should == span(
+        event:      event('app.method', 'MyClass#three'),
+        started_at: 5_000,
+        duration:   1_000)
+
+      t.spans[2].should == span(
+        event:      event('app.winning', 'Win'),
+        started_at: 7_000,
+        duration:   1_000)
+
+      t.spans[3].should == span(
+        event:      event('app.method', 'MyClass.singleton_method'),
+        started_at: 9_000,
+        duration:   1_000)
+
+      # Root span
+      t.spans[4].should == span(
+        event:      event('app.rack.request'),
+        started_at: 0,
+        duration:   10_000,
+        children:   4)
+    end
+
   end
 
 end